requirements.txt
selenium==4.15.2
beautifulsoup4==4.12.2
pandas==2.1.3
undetected-chromedriver==3.5.4
python-dotenv==1.0.0
openpyxl==3.1.2
webdriver-manager==4.0.1
flask==2.3.3
requests==2.31.0
lxml==4.9.3
gunicorn==21.2.0